First Commit Creating a Registration Screen: Development of the Sign Up Screen with 3 UITextFields and 1 Button inside of a Stackview + a photoUploadBtn
Creating a Registration Screen
Instagram Firebase

We'll start off the course by creating a clean and easy layout for our
user to enter their registration information.  In order to achieve
precision with the layout information, I'll show you how to do this
using the programmatic approach of layout anchors.

It's really fun once you get the hang of it so I think you'll enjoy all
the layout techniques in this lesson.