Commit2: UIColor and Anchor Extension
Create an Extension File for UIColor and Anchoring UIView Elements.
Replaced the old anchoring with the new anchor method to place the
elements of the signup view correctly.

Having seen how verbose and time consuming it can be to layout views
with UIKit, let's make our lives easier by creating an extension that
does all the heavy lifting in one single line of code.

Finally we will refactor our layout code to use this extension method
and examine the amount of work we'll save now and moving forward.