Commit3: Create a Firebase Project, Install the SDK, include into Project,  Registration of DummyUser, and activate Listener for SignUpBtn to change color
Firebase SDK and User Registration
Instagram Firebase

Moving on now that we have our registration page setup, we'll go
through the very basic steps of setting up our project to include the
necessary Firebase frameworks.  To be more specific, we'll be
installing Firebase Database, Authentication, and Storage to power our
Instagram application.

Next, we'll go ahead and create a few dummy users with the available
input fields inside of our UI.  After we confirm that this is working,
I'll demonstrate a very useful effect that allows you to show whether
or not the registration form is correctly filled out before allowing
the user to sign up.